Mark empty abouts as read
About changes were always marked as unread, which is fine 9/10 of the time. However due to an incorrect check in `onAboutReceived` we also applied unread to empty abouts. These should not marked as unread.

This includes a flyway (21) to update people's directory to mark all the profiles with empty abouts that were wrongly marked as unread.. As read.