RP styles text improvement
Since we show this text on both the player's dropdown hover AND on other player's choice option.. It made more sense to refer to it as "this character".

This makes sure that we can properly use the string in both cases, because "you" felt a bit weird when you read it on someone else's profile.

Images of this in the attached PR.