Mark empty abouts as read (combined flyway with inactive glances) (#1021)
* Mark empty abouts as read

About changes were always marked as unread, which is fine 9/10 of the time. However due to an incorrect check in `onAboutReceived` we also applied unread to empty abouts. These should not marked as unread.

This includes a flyway (21) to update people's directory to mark all the profiles with empty abouts that were wrongly marked as unread.. As read.

* Handle inactive glance removal in same flyway

* Apply suggestions from code review

Fair point, I was sorta thinking about that one. (Thanks Meorawr!)

Co-authored-by: Daniel Yates <me@meorawr.io>

* Optimize flyway further

* Optimize functions + Flyway

* Bad git, don't revert the actual fix

* Optimize shouldShowTemplate functions

---------

Co-authored-by: Daniel Yates <me@meorawr.io>