Add background browser for about pages
This implements a basic browser to visually select backgrounds for about
pages, replacing the now almost 50 element long dropdown. Most of the
code is just a copy/paste of the icon browser, followed by some of the
worst code ever to integrate its use into all of our about editors.