Only allow disabling purge in debug builds (#1028)
This setting is dangerous for normal users and always leads to data loss
thanks to the constant table limit. If someone wants to manually tweak
the setting after it's removed they still can (and we won't flyway it
away), but for normal users going forward we should try to prevent them
from losing their data.