Tweak and remove comms pooling settings (#1027)
* Tweak and remove comms pooling settings

With the change to per-prefix queues in 10.2.7 our logic for determining
whether or not we should use queue pools for sending profile data has
needed a look at.

Firstly, the settings for controlling this have been entirely removed.
The rationale here is we don't want to give users the power to degrade
comms across their system, and it's hard to change defaults if they're
configurable.

Secondly, in terms of functional changes we now force pooling for all
data transmissions instead of only when a certain profile size is
exceeded, and additionally we now only use 5 queues down from 10.

Forcing transmissions to always go via our 5 queues is mostly a
consistency change. With 10.2.7 the default queue logic employed by
CTL/Chomp is to use a per-prefix queue, so since that patch all profile
data transfers that were under 4kb were effectively running in a serial
manner down one queue. This should allow us to send small profiles
incrementally to multiple players at once again without requiring they
be larger than the 4 KB threshold that previously triggered pooling.

For the queue count, the reduction is to just ensure that transfers to
certain players complete a bit faster. The old choice of 10 queues was
picked based upon default queue logic being per-prefix-and-target, but
now that all addons are per-prefix-only us hogging 10 queues is a bit
unfair to other addons in the environment.

* Remove unused loc strings for removed settings