Use same anchors when bar is and isn't present (#1026)
Currently the new scrollbox-based register list adjusts its anchors so
that the scrollbox frame doesn't overlap the scrollbar.

Visually, the overlap causes no issues if it occurs - and avoiding the
overlap means our "Flags" and a few other columns end up being
positioned too far left of their actual headers when the window is
resized horizontally, so things look weird.

Until this entire UI is blown up and rewritten from scratch, for now
just let the scrollbox always use the same anchors and allow it to
overlap the scrollbar a bit in terms of its rect. This brings the
columns *mostly* back into alignment.