Add option to display full relations in tooltips
This expands the "Show relation" option converting it to an enum and
allows displaying the full relation description in the tooltip if set to
the new "Show full description" value.

The default is "Show name only" which matches the existing behavior of
only displaying a 'Relation: <name>' line in the tooltip.

Flyway code from the boolean setting is not included; we've not done a
release with the new setting, so not much point.