Add basic Player object
Unlike the existing Player class, this object is only intended to be
used as an object-based reference to an OOC player via a "name-realm"
ID. It won't have anything relating to character or profile data made
accessible.

Odds are, this object won't be something that's instantiated too
frequently from actual code - it just exists as a plumbing type to be
used by the Character and Profile objects which will have GetOwner
methods which return Players.