Remove duplicated loc keys + settings dropdown consistency (#1039)
>Loc keys

This will be on-going for a bit, as it isn't that easy to spot duplicates with it being very much a manual review idea.
Plus I've opted to go for the biggest offenders for now, as this list can get pretty exhaustive otherwise.

NOTE: Some have remained as while in English they are duplicates, in some languages they are handled/written differently.

- `CM_DO_NOT_SHOW` now handles `REG_PLAYER_STYLE_HIDE`, `REG_PLAYER_RELATIONSHIP_STATUS_UNKNOWN`, `CO_TOOLTIP_RELATION_LINE_HIDDEN`.
  - It also handles `CO_TOOLTIP_GUILD_HIDDEN`, `CO_TOOLTIP_HEALTH_DISABLED`, `CO_TOOLBAR_VISIBILITY_3`, `NAMEPLATES_CONFIG_UNIT_STATE_HIDE` as they might as well be "Do not show".
- `REG_PLAYER` now handles `AUTOMATION_CATEGORY_CHARACTER`.
- `NAMEPLATES_NAME` now handles `NAMEPLATES_MODULE_NAME`, `NAMEPLATES_CONFIG_MENU_TITLE`, `AUTOMATION_CATEGORY_NAMEPLATES`.
- `REG_PLAYER_CHARACTERISTICS` now handles `REG_PLAYER_CARACT`.
- `REG_TITLE` now handles `REG_PLAYER_TITLE` and `REG_COMPANION_TITLE`.

>Dropdowns

Dropdowns in settings will generally be: Change nothing > Do not show > Change something > Change all (the bigger/more permissive the change, the lower in the list. Where applicable!).

>Settings Tooltips

I never checked the mature dictionary tooltips, they have now been made consistent.

PS: This does not include Extended (for now).