Add an absolute unit of an edit button
This tweaks our "Edit" button for characteristics and about pages to
instead be a big red button with a pencil icon inside of it. The
rationale here is that our "Edit" button doesn't actually fit its text
in certain locales and looks stupid as a result, so let's try something
a bit special.