Place the tooltip on the tab button instead
This means users won't need to hover *specifically* the icon to figure
out what it means - just mousing over the tab will tell them. This in
turn also means that the icon doesn't start intercepting mouse events,
which are a pain to propagate to parents until 11.0.

Additionally - the tooltip body includes the icon before the descriptive
text now to make it a bit clearer what the icon actually means.