Disable TOC splitting in packager
We no longer need the packager to create individual TOC files for each
client version with the comma-delimited Interface support.

Note that Classic users will need to be told to fully exit the game
before updating however, as if they don't then the client will pretend
the addon doesn't exist if the "_Classic.toc"/"_Vanilla.toc" files are
deleted as part of an upgrade.