Add truncation tooltips to most fields
We'll now show tooltips when mousing over the name, title, and PS trait
name fields for profiles if the text is too large to be displayed.

Additionally the styling of truncation tooltips should be more uniform
now, however one exception to this is the name and eye color fields as
these use embedded "|c" markup and so override the color used by the
tooltip line description.