Centralize soundkit definitions and add more sounds
This defines a new global (TRP3_InterfaceSounds) for any common sound
kits that we're using on interactions across the UI. Not all sounds have
entries in this table - if a specific interaction is unique enough to
have its own soundkits, it's free to call PlaySound with the SKID
directly. This is more just a centralization of the *commonly used*
ones, or ones that are otherwise a pain in the rear to manage.

Additionally, some additional sounds are now played upon interactions:

- Dragging glances now plays a sound effect similar to dragging MI/PS
  fields.
- The color picker "Presets" button no longer double-plays a sound.
- The music browser plays a click sound when selecting a song.
- The mature profile editor plays click sounds on various interactions.