Limit maximum open directory pages
This limits the number of maximum directory pages that can be opened at
once, with the goal of preventing users from too easily opening a lot of
profile pages and clogging up their menus unnecessarily.

At the moment this is hardcoded to only keep open the five most recent
pages. We could consider making this a setting if users request it. The
limit is shared across both player and companion profiles.