Convert music browser to use scrollbox list view
This converts the music browser to use a scrollbox based listview
instead of our old crusty and ugly list view logic.

The browser isn't changed too much apart from a small modernization
of the code behind it otherwise, so should remain familiar to
users. One small usability change however is that the search filter now
behaves similarly to the icon browser in that it'll perform searches
with punctuation characters stripped and case insensitively.