Strip consecutive spaces from FT fields (#1092)
* Strip consecutive spaces from FT fields

People keep putting a silly number of spaces in their title fields to
force an effect of having multiple titles split across separate lines.

We've historically discouraged this because it assumes titles are only
ever shown in tooltips and will actually wrap the text; this isn't
guaranteed to the case (FTs sent over MSP to other addons may not be
displayed with wrapping), and in nameplates that's exactly what happens.

As we're adding other tooltip restrictions, let's kill consecutive
spaces in the FT field for player profiles and the TI field for
companion profiles. This is uniformly applied across the characteristics
page views, the tooltip, and nameplates. The stripping occurs before
cropping so as to avoid too eagerly truncating text when it doesn't
strictly require it.

If this proves successful we can look into applying the same change to
other fields potentially.

* Mount tooltips are stupid