Scrollable switch profile dropdowns if list too big
This applies to both the character and companion profile dropdown lists (toolbar and target frame, respectively).

Note that these do not scroll down to the currently selected profile (this is not an option as-is right now). So if your currently selected profile is at the bottom, you may have to scroll down a bit to see it.