Add extra highlight to LoC
Given that it's a more verbose explanation already, I feel we can get away with an extra highlight for people quickly 'hovering' over it and want to know its most obvious examples.

The highlight is also far enough away from the other highlight that it works without much issues, at least in my opinion.

Also updated the baguette version with this change already.