Remove maximize/minimize buttons
Seemingly no one wants these, and we could have just gotten rid of them
when we added the setting to hide these buttons back in 2.5.3.

This change will probably break Extended a bit because it inherits the
"TRP3_MainFrame" template and so expects there to be minimize and
maximize buttons. That it inherits this template is problematic anyway
so I'm not really willing to keep the buttons around on the template
itself - if it wants them back, it can add them.