Kill styled buttons and tweak main frame
The styled button code was required because Classic sucks, however
frankly I'm getting a bit tired of Classic sucking. We'll package and
ship the retail redbuttons asset and Classic can just experience changes
for the first time in its miserable life. As a result, the styled button
template and code is now gone replaced by static templates for each
button style individually. Yay.

For the main frame, to break extended a bit more we're renaming the main
frame template and turning it into an art-only template. This should
make it actually *easier* to use in extended going forward because it'll
never pick up widgets or other incompatible changes that only ever apply
to the core's main frame.