Set bounds on resize buttons
The resize button utility is stupid and doesn't set its own resize
bounds. If you shrink the resize shadow enough, it'll zoom off the
screen comically as a result.

Set the bounds and we can remove a bunch of code that turns the text
red. Also add support for maximum size constraints if they're ever
needed, and get rid of an unnecessary local.