Reset window anchor to TOPLEFT when resizing
Because of our ass approach to resizing there's a few edge cases where
changing the dimensions of the window can result in it not resizing
relative to the top-left corner as one might visually expect. For
example, if the window is super-wide then making it thin again will
often resize relative to the CENTER anchor of the screen.

Prior to resizing (and at some other points) we'll now try and force the
window to anchor relative to the top-left corner of its parent.