Only check world cursor config if TT module is on (#1097)
The tooltip setup code initializes hooks on GTT in order to drive the
internal event for mouseover changes. As part of this, we'd also check
for world cursor units. In 3.0.0 we added a setting to disable the world
cursor units, but this setting is only initialized if the TT module is
actually enabled.

As such - if the module is disabled, we shouldn't query the unregistered
config key.