Rewrite handling of toolbar buttons
Convert toolbar buttons to use a mixin and move all the management of
the button decorations to it, rather than having about five separate
places in the rest of the toolbar code doing ad-hoc updates to parts of
the buttons and making assumptions all the damn time.

For the core layout logic, source toolbar buttons from a pool as well so
we can drop the manual recycling management.

Additionally - drop onMouseDown support for buttons. Everything uses
onClick so let's just force code to go that route rather than keep
leaving ourselves open to "consistency" surprises.

Finally, tooltips are now handled via the new tooltip description logic.
This also adds support for a (presently unused) 'tooltipInstructions'
field on data objects which can be used to avoid all the ugly string
concatenation we've got going on in half the tooltip model handlers.