Rewrite toolbar frame layout logic
Move (and rename) the template and properly convert it to a mixin this
time, unlike the old semi-complete work that was done when adding basic
visibility settings.

The layout logic of the toolbar now uses the GridLayout functionality in
FrameXML. Ideally this would be a GridLayoutFrame but those don't work
in Classic.