De-virtualize the toolbar
There's only ever going to be one toolbar so it doesn't need to be a
template, especially when we forcibly create the frame on initialization
anyway rather than conditioning it upon the module being enabled.

Also take the opportunity to rename it so that we can reclaim
'TRP3_Toolbar' in the global namespace for future work.