Improve handling of zero button case
Old toolbar logic would hide the container when there's no buttons, so
let's also do that. Additionally recalculate the size of the toolbar
frame itself so that it always encompasses both the title and container
sections if shown. This makes the screen clamping and anything anchoring
to the toolbar more accurate.