Convert names to parent keys and hide swatches
Modernize more of the characteristics page code and apply more mixins to
things, as well as clean up the templates.

The templates for lines in the read and edit views no longer require
global names and have proper parentKey attributes defined and
consistently used.

For read mode, we now use the consult pool for a few more regions
covering basic info, MI, and PS. Mixins have been added to hide away
some of the internal details and automatize some of the busywork that
the layout code was previously doing.