Toolbar Modernization (and also IC/OOC icons) (#1108)
* Tweak IC and OOC icons for toolbar

Colorblind feedback indicates that, surprisingly, red and green is not a
great combination. White and black however seems to be universally
adored and looks stylish!

* Fix inability to set tooltip offsets

Defining GetAnchorOffset twice perhaps unsurprisingly does not result in
expected behaviours. Add a SetAnchorWithOffset as the common case if you
want an offset is that you want an anchor as well.

* Allow tooltip instructions to be nil

CreateInstructionTooltip will be a bit easier to use everywhere if it
supports the instruction parameter being nil, rather than assuming it's
always a table.

* Rewrite handling of toolbar buttons

Convert toolbar buttons to use a mixin and move all the management of
the button decorations to it, rather than having about five separate
places in the rest of the toolbar code doing ad-hoc updates to parts of
the buttons and making assumptions all the damn time.

For the core layout logic, source toolbar buttons from a pool as well so
we can drop the manual recycling management.

Additionally - drop onMouseDown support for buttons. Everything uses
onClick so let's just force code to go that route rather than keep
leaving ourselves open to "consistency" surprises.

Finally, tooltips are now handled via the new tooltip description logic.
This also adds support for a (presently unused) 'tooltipInstructions'
field on data objects which can be used to avoid all the ugly string
concatenation we've got going on in half the tooltip model handlers.

* Add tooltipInstructions support to LDB integration

* Split out toolbar code somewhat

Move the button to its own file and also ship the constants for config
keys to a separate one. As there's some shared logic between the main
bar and the buttons, throw in a utils file as well.

* Rewrite toolbar frame layout logic

Move (and rename) the template and properly convert it to a mixin this
time, unlike the old semi-complete work that was done when adding basic
visibility settings.

The layout logic of the toolbar now uses the GridLayout functionality in
FrameXML. Ideally this would be a GridLayoutFrame but those don't work
in Classic.

* Build toolbar after init

It works either way for now, but it's stupid to build and _then_
initialize.

* ElvUI should probably skin the right frame

* Updating buttons should in fact update them

Would rather not have it such that updating button potentially
_replaces_ the currently assigned element data, so instead just discard
the second parameter and tell the button to update from its existing
view of the data.

Additionally, clean up nonsense elsewhere. Buttons update automatically,
so there's no point in having additional onUpdate functions that update
the button twice or refresh tooltips that don't exist.

* De-virtualize the toolbar

There's only ever going to be one toolbar so it doesn't need to be a
template, especially when we forcibly create the frame on initialization
anyway rather than conditioning it upon the module being enabled.

Also take the opportunity to rename it so that we can reclaim
'TRP3_Toolbar' in the global namespace for future work.

* Don't register config pages if not enabled

With the splitting of the settings it seems the config pages for both
the toolbar and the target frame were unconditionally registered even if
their respective modules were disabled and thus would have no settings.

* Improve handling of zero button case

Old toolbar logic would hide the container when there's no buttons, so
let's also do that. Additionally recalculate the size of the toolbar
frame itself so that it always encompasses both the title and container
sections if shown. This makes the screen clamping and anything anchoring
to the toolbar more accurate.

* Hide is not show

Who even made this setting

* Usual parameter misordering in split