Small tweaks to CO/CU debouncing on Misc page
Ensure that we only process changes to the fields that are initiated via
user input rather than both input and SetText calls. This *may* resolve
an edge case where the CO/CU fields from other profiles could be copied
to the user's own profile.

Additionally, lower the debounce timeout to 0.2s for this page
specifically. With 0.5s it's possible in some cases to type input, open
another profile, and lose your input. Ideally we'd flush the input when
the page is hidden but that's more effort than this code deserves, so
tightening it to 0.2s feels a bit more sane.