Goodbye VA
Kill support entirely for the VA field; there's no real point keeping it
around and odds are no one's importing profiles from 2018. We can't test
such imports as we don't have them ourselves, and even if we did there's
bound to be other issues that will occur because of all the oter flyways
that touch profiles.

If someone does import an ancient profile - it won't error, but the
values will all display as their defaults until the user goes in and
edits them. So not the end of the world at least.