Add new HTML container templates
This implements two new self-contained templates for displaying HTML
content with support for hyperlink handlers and an updated link copy
dialog.

The 'TRP3_HTMLFrameTemplate' is a standard frame template that contains
a SimpleHTML child. The template exposes most methods that normal
SimpleHTML does, with the addition of a few extras for more conveniently
handling the translation from our profile markup to HTML
("SetRichText"). The template handles resizing itself to the bounds of
the displayed content automatically.

The 'TRP3_HTMLScrollFrameTemplate' is similar to the previous but it
wraps everything inside a ScrollBox container and additionally provides
a built-in scrollbar with consistent positioning, as well as preset
padding.