Integrate new HTML templates into dashboard
This converts the dashboard to use the new scrolling HTML template for
its tab views. Additionally, the amount of code for the dashboard has
been cut down significantly.

The changelog has been slightly compressed to reduce spacing and play a
bit nicer with the new default font set, and support has been added for
displaying list markup with bullets.