Remove transmission support for PS VAs (#1114)
* Remove transmission support for PS VAs

This removes support for sending the old 0-6 range VA field as part of
trait comms. The VA field was deprecated and replaced by V2 all the way
back in 2018, so there should be nothing on the network that requires
it.

For receipt, we still process the VA field if there's no V2 field in
viewed profile data. The cost of keeping this around is minimal for now,
and we can kill it off some other time if that ever changes.

Additionally some improvements have been made to our serialization of
custom trait colors for transmission to try and reduce wire consumption.

Unfortunately when custom colors were added they were transmitted in RGB
table form rather than hex colors, so are subject to a bit of blow-up if
color components are a bit too exact. We now try and round them to 2
significant digits prior to transmission, so there's a bit of lost
information but it should be mostly accurate.

* Goodbye VA

Kill support entirely for the VA field; there's no real point keeping it
around and odds are no one's importing profiles from 2018. We can't test
such imports as we don't have them ourselves, and even if we did there's
bound to be other issues that will occur because of all the oter flyways
that touch profiles.

If someone does import an ancient profile - it won't error, but the
values will all display as their defaults until the user goes in and
edits them. So not the end of the world at least.

* Remove old constants for VA