Reduce default size of color swatch
The 16x16 size of the swatch was selected while messing around with
getting the pixel util calls mostly-correct to avoid random 2px edges.

For characteristics the swatch size doesn't need to be so large, so
align with Blizzard's own default size for their template and go down to
14x14 instead.

For the color picker dropdown, that explicitly sets the size to 16x16
as we need the color to be very clearly shown there.