Fix hiding of icons and titles on Kui nameplates
The Kui nameplate logic determines if it "can" customize a nameplate
based upon both upon our internal requirements for its state (ie. is it
initialized, does it have a unit, etc.) but also based upon whether or
not it's a personal nameplate or a nameplate for a unit that has a valid
reaction.

The problem is that this combined logic results in us not triggering the
name text update for nameplates that have transitioned from "can-update"
to "cannot-update". This results in the icon and full title updates
never being processed, as these are invoked from within the name text
update hook.

This commit splits the "can" logic into "can" and "should", with most
functions converted to checking "should". When updating a nameplate if
we determine that we shouldn't update it, we'll forcefully update the
icon and full titles anyway in order to properly hide them.