Fix tooltips for party units on other maps
The tooltip logic when the mouseover unit changes explicitly consults
the "mouseover" unit or falls back to the world cursor support in
mainline when enabled.

For units on different maps to the current player, the client never
associates the "mouseover" unit to them - it'll instead display the
tooltip with their party/raid unit token which we've never checked.

The current tooltip unit logic has been adjusted to consult the default
GTT if shown and use its unit if set, before then falling back to the
"mouseover" token and world cursor support.