Flag HTMLFrameTemplate as being a layout frame
As the HTML frame template effectively behaves as a ResizeLayoutFrame,
we can flag it as being a layout frame-ish with similar characteristics.

The benefit this provides is that if a HTML frame is used as child of a
layout frame, it'll automatically have its Layout method called in
response to a parent updating its own layout.