Dig us out of more formatting holes
We also need to resolve similar errors when formatting the description
for display. Generalize the description formatting logic to just look
for the exact tokens we want, and avoid string.format entirely.

Ideally we would have escaped % inputs entirely prior to releasing the
feature, but it's a bit late to rewind now.