Adjust background for characteristics contrast (#1122)
The background color was sampled from a bit closer to the edge of the
characteristics panel, whereas text itself generally ends up toward this
lighter patch of the parchment background.

Selecting this lighter patch as a background gives us some better
constrast at the 'low' default level. As usual 'medium low' continues to
look the best, so we might want to consider forcing that as the minimum
level for this page in general or adjusting our default somewhere down
the line.