Kindly resolve relation formatting quirks (#1137)
* Don't blindly format descriptions in relations UI

If an intrepid user sticks a lone '%' into a relation description or
otherwise gives us a format string that would be invalid, we currently
catastrophically blow up and kill the entire registration of the
relations UI.

When formatting a description for editing we only need to care about
translating the persisted tokens (%1$s/%2$s) to %p/%t, so we can avoid
blowing up by not formatting things here and just using a gsub.

* Dig us out of more formatting holes

We also need to resolve similar errors when formatting the description
for display. Generalize the description formatting logic to just look
for the exact tokens we want, and avoid string.format entirely.

Ideally we would have escaped % inputs entirely prior to releasing the
feature, but it's a bit late to rewind now.