Improve resizing behavior (#1141)
* Perform resizing on mouse down rather than drag

Making small adjustments to sizes is a pain with drag start/stop events
as one needs to move the cursor a small amount before those handlers
fire. As we're setting the cursor on mouse interaction anyway, let's
just use that script handler pair to drive the whole damn thing.

* Set bounds on resize buttons

The resize button utility is stupid and doesn't set its own resize
bounds. If you shrink the resize shadow enough, it'll zoom off the
screen comically as a result.

Set the bounds and we can remove a bunch of code that turns the text
red. Also add support for maximum size constraints if they're ever
needed, and get rid of an unnecessary local.

* Reset window anchor to TOPLEFT when resizing

Because of our ass approach to resizing there's a few edge cases where
changing the dimensions of the window can result in it not resizing
relative to the top-left corner as one might visually expect. For
example, if the window is super-wide then making it thin again will
often resize relative to the CENTER anchor of the screen.

Prior to resizing (and at some other points) we'll now try and force the
window to anchor relative to the top-left corner of its parent.

* These keys aren't needed

* Fix merge damage