Scrolling HTML rework (#1129)
* Add new HTML container templates

This implements two new self-contained templates for displaying HTML
content with support for hyperlink handlers and an updated link copy
dialog.

The 'TRP3_HTMLFrameTemplate' is a standard frame template that contains
a SimpleHTML child. The template exposes most methods that normal
SimpleHTML does, with the addition of a few extras for more conveniently
handling the translation from our profile markup to HTML
("SetRichText"). The template handles resizing itself to the bounds of
the displayed content automatically.

The 'TRP3_HTMLScrollFrameTemplate' is similar to the previous but it
wraps everything inside a ScrollBox container and additionally provides
a built-in scrollbar with consistent positioning, as well as preset
padding.

* Use consistent markup for 'More modules'

The other tabs in our dashboard use h1 for headers, but these use h2.
Make them h1.

* Integrate new HTML templates into dashboard

This converts the dashboard to use the new scrolling HTML template for
its tab views. Additionally, the amount of code for the dashboard has
been cut down significantly.

The changelog has been slightly compressed to reduce spacing and play a
bit nicer with the new default font set, and support has been added for
displaying list markup with bullets.

* Reset scroll position on text changes

Additionally, add passthrough parameters for disabling interpolation
when calling scroll related APIs.

* Factor out text updates to separate method

* Rename padding keys for scroll frame template

The 'initial' prefix isn't really that consistent anyway, as some of the
others are also 'initial' only.

* Flag HTMLFrameTemplate as being a layout frame

As the HTML frame template effectively behaves as a ResizeLayoutFrame,
we can flag it as being a layout frame-ish with similar characteristics.

The benefit this provides is that if a HTML frame is used as child of a
layout frame, it'll automatically have its Layout method called in
response to a parent updating its own layout.

* We shorten too