Strictly validate interface icons on load
If someone installs a Mainline icon pack in Classic then the interface
icon selector can get a bit confused.

What'll happen is it'll see that the preferred mainline icons for each
part of the interface exist, but use of them will fail in conjunction
with our fancy new BorderedIconTemplate system because that uses LRPM
to strictly validate icons against a defined set that we're actually
expecting to find in Classic.

As the use of Mainline icons in Classic is wholly unsupported, let's
just apply this same stricter LRPM-based check earlier and completely
remove the possibility that Mainline icons can be picked up from custom
icon packs.