Add Group and Category TOC directives (#1160)
* Add Group and Category TOC directives

Initial translations for these categories are machine-translated, but
they look consistent with Blizzard's own strings as used for realm
category names with the exception of ptBR; however the ptBR translation
is much longer so it might be the case that Blizzard specifically
chose a much shorter string so as to avoid UI issues.

Explicit Groups have been defined on both TOCs to prevent any cases
where the auto-grouping might for some reason decide to break our
addons apart, eg. if someone made an addon called "tot" and made it
depend upon "totalRP3".

* Use wiki translation for ruRU category

This one was supplied by a native speaker supposedly, so trusting it.

* Update ptBR translation to match wiki

* Align with wiki (and Blizzard)