Remove ClearAllPoints call after we show the tooltip (#1197)
Blizzard are changing ClearAllPoints in 11.2.0 to invalidate frame rects. To accommodate, we need to remove this post-Show ClearAllPoints call or we'll end up with an invisible tooltip.

```lua
-- NOTE FOR ADDON DEVELOPERS:
-- Starting this patch, region:ClearAllPoints() will immediately invalidate the rect.
-- This means you cannot rely on calling GetWidth, GetHeight, GetTop/Left/Bottom/Right, or GetRect after ClearAllPoints.
-- Any measurement calculations relying on the previous rect should occur before calling ClearAllPoints.
```

God knows why the code was doing this all along. Will probably break something like fading, but that's better than literally _not having a tooltip whatsoever_.